# docs/openapi.yaml
{
    "openapi": "3.2.0",
    "info": {
        "title": "Znamky API",
        "description": "JSON API for Znamky.\n\nAuthenticate with `POST /api/login` (verified users only), then send\n`Authorization: Bearer <token>`. Refresh via `POST /api/token/refresh`.\n\nValidation errors for auth request DTOs return **422** with a\n`ValidationFailed` body (`title`, `detail`, `violations[]`).\n\nGraphQL: `POST /api/graphql` (public entrypoint; auth mutations plus\nJWT-protected stamp catalog queries), IDE at `/api/graphql/graphiql`.\n\nHuman docs: `docs/auth.md` (auth), `docs/stamps.md` (stamp catalog).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": ""
        }
    ],
    "paths": {
        "/api/forgot-password": {
            "post": {
                "operationId": "api_forgot_password_post",
                "tags": [
                    "Authentication"
                ],
                "responses": {
                    "204": {
                        "description": "Always returned; does not reveal whether the email exists"
                    },
                    "422": {
                        "description": "Validation failed (invalid email)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationFailed"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded (Retry-After header set)"
                    }
                },
                "summary": "Request password reset email",
                "description": "If the email belongs to a user, sends a reset token by email. Optional locale (en default, cs) selects the email language. Rate limited per client IP.",
                "requestBody": {
                    "description": "Account email and optional email locale",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "locale": {
                                        "type": "string",
                                        "enum": [
                                            "en",
                                            "cs"
                                        ],
                                        "default": "en",
                                        "description": "Language for the password-reset email"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "security": []
            }
        },
        "/api/logout": {
            "post": {
                "operationId": "api_logout_post",
                "tags": [
                    "Authentication"
                ],
                "responses": {
                    "204": {
                        "description": "Logged out"
                    },
                    "401": {
                        "description": "Missing or invalid JWT"
                    }
                },
                "summary": "Logout",
                "description": "Blocklists the current JWT. If `refresh_token` is sent, that refresh token is revoked; otherwise all refresh tokens for the user are revoked.",
                "requestBody": {
                    "description": "Optional refresh token to revoke (omit to revoke all for the user)",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "refresh_token": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                }
                            }
                        }
                    },
                    "required": false
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "operationId": "api_me_get",
                "tags": [
                    "Authentication"
                ],
                "responses": {
                    "200": {
                        "description": "Current user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "id",
                                        "email",
                                        "roles"
                                    ],
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "roles": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT"
                    }
                },
                "summary": "Current user",
                "description": "Returns the authenticated user profile.",
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "operationId": "api_register_post",
                "tags": [
                    "Authentication"
                ],
                "responses": {
                    "201": {
                        "description": "User created (email not verified yet)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "id",
                                        "email"
                                    ],
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed (invalid or duplicate email)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationFailed"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded (Retry-After header set)"
                    }
                },
                "summary": "Register a new user",
                "description": "Creates an unverified user and sends a verification token by email. Optional locale (en default, cs) selects the email language. Re-registering an unverified email rotates and resends the token. Verified emails return 422. The account password is set later via verify-email. Rate limited per client IP.",
                "requestBody": {
                    "description": "Registration email and optional email locale",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "locale": {
                                        "type": "string",
                                        "enum": [
                                            "en",
                                            "cs"
                                        ],
                                        "default": "en",
                                        "description": "Language for the verification email"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "security": []
            }
        },
        "/api/reset-password": {
            "post": {
                "operationId": "api_reset_password_post",
                "tags": [
                    "Authentication"
                ],
                "responses": {
                    "204": {
                        "description": "Password updated"
                    },
                    "422": {
                        "description": "Invalid payload, or invalid/expired token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationFailed"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded (Retry-After header set)"
                    }
                },
                "summary": "Reset password with email token",
                "description": "Sets a new password using the raw token from the reset email. Revokes all refresh tokens for the user. Does not verify the email; an unverified account must still confirm the original verification token. Rate limited per client IP.",
                "requestBody": {
                    "description": "Reset token and new password",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "token",
                                    "password"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "security": []
            }
        },
        "/api/verify-email": {
            "post": {
                "operationId": "api_verify_email_post",
                "tags": [
                    "Authentication"
                ],
                "responses": {
                    "204": {
                        "description": "Email verified and password set"
                    },
                    "422": {
                        "description": "Invalid payload, or invalid/expired token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationFailed"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded (Retry-After header set)"
                    }
                },
                "summary": "Verify email address",
                "description": "Confirms the account using the raw token from the verification email and sets the account password. Rate limited per client IP.",
                "requestBody": {
                    "description": "Verification token and the password chosen by the mailbox owner",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "token",
                                    "password"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "security": []
            }
        },
        "/api/login": {
            "post": {
                "operationId": "login_check_post",
                "tags": [
                    "Login Check"
                ],
                "responses": {
                    "200": {
                        "description": "User token created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "readOnly": true,
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "refresh_token": {
                                            "readOnly": true,
                                            "type": "string",
                                            "nullable": false
                                        }
                                    },
                                    "required": [
                                        "token"
                                    ]
                                }
                            }
                        }
                    }
                },
                "summary": "Creates a user token.",
                "description": "Creates a user token.",
                "requestBody": {
                    "description": "The login data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "password": {
                                        "type": "string",
                                        "nullable": false
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/api/token/refresh": {
            "post": {
                "operationId": "gesdinet_jwt_refresh_token_post",
                "tags": [
                    "Login Check"
                ],
                "responses": {
                    "200": {
                        "description": "JWT refreshed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "readOnly": true,
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "refresh_token": {
                                            "readOnly": true,
                                            "type": "string",
                                            "nullable": false
                                        }
                                    },
                                    "required": [
                                        "token"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The refresh token is missing, unknown or has expired"
                    }
                },
                "summary": "Exchanges a refresh token for a new JWT.",
                "description": "Returns a new JWT, and a refresh token to use for the next exchange.",
                "requestBody": {
                    "description": "The refresh token issued alongside the JWT being replaced",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "refresh_token": {
                                        "type": "string",
                                        "nullable": false
                                    }
                                },
                                "required": [
                                    "refresh_token"
                                ]
                            }
                        }
                    },
                    "required": true
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationFailed": {
                "type": "object",
                "required": [
                    "title",
                    "detail",
                    "violations"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Validation Failed"
                    },
                    "detail": {
                        "type": "string",
                        "example": "The given data failed validation."
                    },
                    "violations": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "required": [
                                "propertyPath",
                                "message"
                            ],
                            "properties": {
                                "propertyPath": {
                                    "type": "string",
                                    "example": "email"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "This value is not a valid email address."
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [],
...
}